pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spoon-toolkits"
version = "0.2.5"
authors = [
  { name="SpoonAI Team", email="team@spoonai.com" },
]
description = "Comprehensive blockchain and cryptocurrency tools for SpoonAI"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # Core dependencies for spoon-toolkits functionality
    "requests>=2.32.3",
    "aiohttp>=3.13.2",
    "pycares>=4.9,<5.0",
    "fastmcp>=2.12.0",
    "pandas>=2.2.0",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.12.3",
    "markdownify>=0.11.6",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "typing-extensions>=4.12.2",
    "nest-asyncio>=1.6.0",

    # Blockchain SDKs and cryptography
    "base58>=2.1.1",
    "PyNaCl>=1.5.0",
    "web3==7.11.0",
    "solana>=0.33.0",
    "solders>=0.19.0",
    "neo-mamba>=3.0.1,<3.2.0; python_version < '3.14'",
    "neo-mamba>=3.2.0; python_version == '3.14'",
    # Crypto PowerData dependencies
    "ccxt>=4.0.0",
    "numpy>=1.26.0",
    "TA-Lib>=0.6.5",
    "asyncio-throttle>=1.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.115.7",
    "uvicorn[standard]>=0.32.0",

    # Data & storage integrations
    "boto3==1.35.99",
    "desearch-py>=1.0.0",

    # Machine learning utilities
    "scikit-learn>=1.4.2",

    # Github GraphQL dependencies
    "gql>=3.5.0",

    # Audio/Voice AI SDKs
    "elevenlabs>=1.0.0",

    # Note: spoon-ai-sdk is required but not listed here to avoid circular dependency
    # Install spoon-ai-sdk first: pip install spoon-ai-sdk
    # Then install spoon-toolkits: pip install spoon-toolkits
]

[project.urls]
"Homepage" = "https://github.com/XSpoonAi/spoon-toolkits"
"Bug Tracker" = "https://github.com/XSpoonAi/spoon-toolkits/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["spoon_toolkits*"]
exclude = ["tests*", "test*"]
